# perf: remove heap allocation in parse_query