--- a/url/benches/parse_url.rs
+++ b/url/benches/parse_url.rs
@@ -19,6 +19,13 @@ fn long(bench: &mut Bencher) {
     bench.iter(|| black_box(url).parse::<Url>().unwrap());
 }
 
+fn long_with_fragment(bench: &mut Bencher) {
+    let url = "https://example.com/parkbench?tre=es&st=uff#fragment";
+
+    bench.bytes = url.len() as u64;
+    bench.iter(|| black_box(url).parse::<Url>().unwrap());
+}
+
 fn plain(bench: &mut Bencher) {
     let url = "https://example.com/";
 
@@ -86,6 +93,7 @@ benchmark_group!(
     benches,
     short,
     long,
+    long_with_fragment,
     plain,
     hyphen,
     leading_digit,