Also updates parse_fragment to work similarly to how parse_query works now where it operates on longer string slices for percent decoding, which helps to not have to do that for every character.

Before:

```
test long           ... bench:         302 ns/iter (+/- 42) = 142 MB/s
test fragment       ... bench:         370 ns/iter (+/- 9) = 140 MB/s
```

After:

```
test long           ... bench:         243 ns/iter (+/- 7) = 176 MB/s
test fragment       ... bench:         275 ns/iter (+/- 9) = 189 MB/s
```